% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{eight_pillars}
\alias{eight_pillars}
\title{Score a stock on eight criteria}
\usage{
eight_pillars(symbol)
}
\arguments{
\item{symbol}{symbol}
}
\value{
list
}
\description{
Score a stock on eight criteria
}
