% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euronext.R
\name{euronext_bonds}
\alias{euronext_bonds}
\title{Get Euronext bond tickers}
\usage{
euronext_bonds(
  mics = c("ALXB", "ALXL", "ALXP", "XPAR", "XAMS", "XBRU", "XLIS", "XMLI", "MLXB",
    "ENXB", "ENXL", "TNLA", "TNLB", "XLDN", "XHFT", "VPXB", "XOSL", "XOAM")
)
}
\arguments{
\item{mics}{MIC codes}
}
\value{
tibble containing symbols and some metrics
}
\description{
This gets all Euronext bond ticker symbols from the Euronext website. Possible MIC codes
are ALXB, ALXL, ALXP, XPAR, XAMS, XBRU, XLIS, XMLI, MLXB, ENXB, ENXL, TNLA, TNLB, XLDN,
XHFT, VPXB, XOSL, XOAM.
}
\examples{
euronext_bonds("XBRU")
}
