% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euronext.R
\name{euronext_tickers}
\alias{euronext_tickers}
\title{Get Euronext tickers}
\usage{
euronext_tickers(
  mics = c("ALXB", "ALXL", "ALXP", "XPAR", "XAMS", "XBRU", "XLIS", "XMLI", "MLXB",
    "ENXB", "ENXL", "TNLA", "TNLB", "XLDN", "XESM", "XMSM", "XATL", "VPXB", "XOSL",
    "XOAS", "MERK")
)
}
\arguments{
\item{mics}{MIC codes}
}
\value{
tibble containing symbols and some metrics
}
\description{
This gets all Euronext ticker symbols from the Euronext website. Possible MIC codes
are ALXB, ALXL, ALXP, XPAR, XAMS, XBRU, XLIS, XMLI, MLXB, ENXB, ENXL, TNLA, TNLB,
XLDN, XESM, XMSM, XATL, VPXB, XOSL, XOAS, MERK.
}
\examples{
euronext_tickers("XBRU")
}
