% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftse.R
\name{ftse_350_tickers}
\alias{ftse_350_tickers}
\title{Get FTSE 350 tickers}
\usage{
ftse_350_tickers()
}
\value{
tibble
}
\description{
This gets all FTSE 350 ticker symbols from Wikipedia.
}
\examples{
ftse_350_tickers()
}
