% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTests.R
\name{jarqueBera}
\alias{jarqueBera}
\title{Perform the Jarque-Bera test for normality.}
\usage{
jarqueBera(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, indicating whether missing values should be removed.
Default is \code{FALSE}.}
}
\value{
A list with components:
  \describe{
    \item{statistic}{The Jarque-Bera test statistic.}
    \item{p.value}{The p-value for the test.}
  }
}
\description{
This function calculates the Jarque-Bera test statistic and corresponding
p-value for assessing the normality of a numeric vector. It can handle
missing values if \code{na.rm} is set to \code{TRUE}.
}
\examples{
# Perform Jarque-Bera test with the default options
jarqueBera(c(1, 2, 3, 4, 5))

# Perform Jarque-Bera test with missing values removed
jarqueBera(c(1, 2, NA, 4, 5), na.rm = TRUE)

}
\seealso{
\code{\link{skewness}}, \code{\link{kurtosis}}
}
