% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformationFuns.R
\name{logReturns}
\alias{logReturns}
\title{Calculate logarithmic returns of a numeric vector.}
\usage{
logReturns(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, indicating whether missing values should be removed.
Default is \code{FALSE}.}
}
\value{
A numeric vector of logarithmic returns.
}
\description{
This function computes the logarithmic returns of a numeric vector by taking
the difference of the logarithm of each element. It can handle missing values
if \code{na.rm} is set to \code{TRUE}.
}
\examples{
\dontrun{
# Calculate logarithmic returns with the default options
logReturns(c(1, 2, 3, 4, 5))

# Calculate logarithmic returns with missing values removed
logReturns(c(1, 2, NA, 4, 5), na.rm = TRUE)

}
}
