% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq.R
\name{nasdaq_tickers}
\alias{nasdaq_tickers}
\title{Get Nasdaq tickers}
\usage{
nasdaq_tickers(exchange = "nasdaq")
}
\arguments{
\item{exchange}{exchange ("nasdaq" or "nyse")}
}
\value{
data frame
}
\description{
This gets all Nasdaq ticker symbols from the Nasdaq website.
}
\examples{
nasdaq_tickers()
}
