
\name{normalize}
\alias{normalize}
\docType{package}
\title{
Generate a normal distribution of an xts columns
}
\description{
Given an xts object x, regenerate columns of x following a normal distribution based on each columns mean and standards deviation
}
\usage{
normalize(x)
}
\arguments{
\item{x}{ an xts like object}
}
\examples{

\dontrun{
normalize()
}
}
}
\author{Joao Claudio Macosso}
