% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq.R
\name{nyse_tickers}
\alias{nyse_tickers}
\title{Get NYSE tickers}
\usage{
nyse_tickers()
}
\value{
data frame
}
\description{
This gets all NYSE ticker symbols from the Nasdaq website.
}
\examples{
nyse_tickers()
}
