% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{packages.github}
\alias{packages.github}
\title{List of packages required from GitHub}
\format{One character vector}
\usage{
data(packages)
}
\description{
List of packages required from GitHub
}
\keyword{datasets}
