% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformationFuns.R
\name{pctReturns}
\alias{pctReturns}
\title{Calculate percentage returns of a numeric vector.}
\usage{
pctReturns(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, indicating whether missing values should be removed.
Default is \code{FALSE}.}
}
\value{
A numeric vector of percentage returns.
}
\description{
This function computes the percentage returns of a numeric vector by taking
the difference between each element and its lagged value, divided by the
lagged value. It can handle missing values if \code{na.rm} is set to
\code{TRUE}.
}
\examples{
# Calculate percentage returns with the default options
pctReturns(c(1, 2, 3, 4, 5))

# Calculate percentage returns with missing values removed
pctReturns(c(1, 2, NA, 4, 5), na.rm = TRUE)

}
