% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\title{Bland-Altman plot comparing the agreement between the LCS and the reference.}
\usage{
plot_bland_altman(data, lcs_column = "lcs", reference_column = "reference")
}
\arguments{
\item{data}{Data frame containing the low-cost sensor and reference measurements.}

\item{lcs_column}{The name of the column holding the LCS measurements.}

\item{reference_column}{The name of the column holding the reference measurements.}
}
\value{
A plot stored as a \code{ggplot} object. If not assigned then prints the plot.
}
\description{
Bland-Altman plot comparing the agreement between the LCS and the reference.
}
\examples{
# Create dummy data frame
n_vals <- 100
df <- data.frame(timestamp=seq.POSIXt(from=as.POSIXct("2020-09-24 09:00:00"),
                                      to=as.POSIXct("2020-09-30 09:00:00"),
                                      length.out=n_vals),
                 ref=rnorm(n_vals, 30, 5),
                 lowcost=rnorm(n_vals, 32, 8))

plot_bland_altman(df, lcs_column="lowcost", reference_column="ref")
}
\seealso{
Other plots: 
\code{\link{plot_reu}()},
\code{\link{plot_scatter}()},
\code{\link{plot_time_series}()}
}
\concept{plots}
