% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_reu}
\alias{plot_reu}
\title{Plot of Relative Expanded Uncertainty (REU) against reference concentrations.}
\usage{
plot_reu(
  data,
  lcs_column = "lcs",
  reference_column = "reference",
  LV = NULL,
  DQO = NULL,
  y_limits = c(0, 200),
  smooth = TRUE
)
}
\arguments{
\item{data}{Data frame containing the low-cost sensor and reference measurements.}

\item{lcs_column}{The name of the column holding the LCS measurements.}

\item{reference_column}{The name of the column holding the reference measurements.}

\item{LV}{The Limit Value for the species being plotted. If provided then a vertical line is drawn in red at this concentration.}

\item{DQO}{The relative uncertainty of the Data Quality Objective for this species. If provided then a horizontal line in blue is drawn at this relative uncertainty.}

\item{y_limits}{Particularly at low concentration values, REU curves can have large outliers that detract from the main trends.
This option allows for the specification of the y-axis limits to zoom in to the region of interest. If \code{NULL} then no limits are applied.}

\item{smooth}{Whether to add a non-linear smooth to summarise the relationship between REU and concentration.}
}
\value{
A plot stored as a \code{ggplot} object. If not assigned then prints the plot.
}
\description{
Plot of Relative Expanded Uncertainty (REU) against reference concentrations.
}
\examples{
# Create dummy data frame
n_vals <- 100
df <- data.frame(timestamp=seq.POSIXt(from=as.POSIXct("2020-09-24 09:00:00"),
                                      to=as.POSIXct("2020-09-30 09:00:00"),
                                      length.out=n_vals),
                 ref=rnorm(n_vals, 30, 5),
                 lowcost=rnorm(n_vals, 32, 8))

# Remove the y-axis limits to show the large relative uncertainity of this imaginary device
plot_reu(df, lcs_column="lowcost", reference_column="ref", y_limits=NULL)

# Add in the LV and DQO
plot_reu(df, lcs_column="lowcost", reference_column="ref", y_limits=NULL, LV=50, DQO=40)
}
\seealso{
Other plots: 
\code{\link{plot_bland_altman}()},
\code{\link{plot_scatter}()},
\code{\link{plot_time_series}()}
}
\concept{plots}
