% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plots the LCS and reference time-series}
\usage{
plot_time_series(
  data,
  lcs_column = "lcs",
  reference_column = "reference",
  time_column = "time"
)
}
\arguments{
\item{data}{Data frame containing the low-cost sensor and reference measurements.}

\item{lcs_column}{The name of the column holding the LCS measurements.}

\item{reference_column}{The name of the column holding the reference measurements.}

\item{time_column}{The name of the column holding the time-stamps, which must be stored as datetime objects.}
}
\value{
A plot stored as a \code{ggplot} object. If not assigned then prints the plot.
}
\description{
Plots the LCS and reference time-series
}
\examples{
# Create dummy data frame
n_vals <- 100
df <- data.frame(timestamp=seq.POSIXt(from=as.POSIXct("2020-09-24 09:00:00"),
                                      to=as.POSIXct("2020-09-30 09:00:00"),
                                      length.out=n_vals),
                 ref=rnorm(n_vals, 30, 5),
                 lowcost=rnorm(n_vals, 32, 8))

# Plot time-series
plot_time_series(df, lcs_column="lowcost", reference_column="ref", time_column="timestamp")
}
\seealso{
Other plots: 
\code{\link{plot_bland_altman}()},
\code{\link{plot_reu}()},
\code{\link{plot_scatter}()}
}
\concept{plots}
