% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{plot_yahoo_financials}
\alias{plot_yahoo_financials}
\title{Plot key financials from Yahoo Finance}
\usage{
plot_yahoo_financials(financials)
}
\arguments{
\item{financials}{output from yahoo_financials()}
}
\value{
plot
}
\description{
Plot key financials from Yahoo Finance
}
