% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantR.R
\docType{package}
\name{quantR-package}
\alias{quantR}
\alias{quantR-package}
\title{Become a Quant with R - Tutorial Codes}
\description{
Open any tutorial code from "Become a Quant with R". Furthermore update the tutorial contents anytime without updating the package.
}
\references{
\url{http://www.quant-r.com/}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.quant-r.com/}
}

}
\author{
Ronald Hochreiter, \email{ron@hochreiter.net}
}
