% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reu.R
\name{reu}
\alias{reu}
\title{Calculates the Relative Expanded Uncertainty of low-cost sensor measurements when compared with a reference instrument.}
\usage{
reu(ref, lcs, k = 2, u_ref = 0)
}
\arguments{
\item{ref}{Reference measurements}

\item{lcs}{Low-cost-sensor measurements}

\item{k}{Multiplier of standard error in relative uncertainty to obtain the expanded uncertainty. Defaults to 2 standard deviations.}

\item{u_ref}{Relative uncertainty of the reference instrument}
}
\value{
A vector of the same length as \code{ref} and \code{lcs} containing the corresponding REU estimates.
}
\description{
Calculates the Relative Expanded Uncertainty of low-cost sensor measurements when compared with a reference instrument.
}
\examples{
# Create dummy data frame
n_vals <- 100
df <- data.frame(timestamp=seq.POSIXt(from=as.POSIXct("2020-09-24 09:00:00"),
                                      to=as.POSIXct("2020-09-30 09:00:00"),
                                      length.out=n_vals),
                 ref=rnorm(n_vals, 30, 5),
                 lowcost=rnorm(n_vals, 32, 8))

reu(df$ref, df$lowcost)

}
