% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{tutorial}
\alias{tutorial}
\title{Open a specific quant-r.com tutorial code}
\usage{
tutorial(tutorial = 0)
}
\arguments{
\item{tutorial}{id of the tutorial to open}
}
\description{
\code{tutorial} opens the R script of a given tutorial code within RStudio
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
