% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xetra.R
\name{xetra_tickers}
\alias{xetra_tickers}
\title{Get Xetra tickers}
\usage{
xetra_tickers()
}
\value{
tibble
}
\description{
This gets all Xetra ticker symbols from the Xetra website.
}
\examples{
xetra_tickers()
}
