% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{yahoo_financials}
\alias{yahoo_financials}
\title{Get Yahoo Finance financials (income statement, balance sheet, cash flow)}
\usage{
yahoo_financials(symbols, reporting = "annual", verbose = FALSE)
}
\arguments{
\item{symbols}{ticker symbols}

\item{reporting}{reporting period ("annual" or "quarterly")}
}
\value{
tibble
}
\description{
Get Yahoo Finance financials (income statement, balance sheet, cash flow)
}
\examples{
yahoo_financials(c("AAPL", "MSFT"), "quarterly")
}
