% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{yahoo_financials_web}
\alias{yahoo_financials_web}
\title{Get Yahoo Finance financials from web page}
\usage{
yahoo_financials_web(symbol)
}
\arguments{
\item{symbol}{ticker symbol}
}
\value{
tibble
}
\description{
Get Yahoo Finance financials from web page
}
