% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{yahoo_history}
\alias{yahoo_history}
\title{Get historical prices from Yahoo Finance}
\usage{
yahoo_history(symbol, interval = "1d", days = 30, end_date = Sys.time())
}
\arguments{
\item{interval}{interval ("1d", "1wk", or "1mo")}

\item{days}{how many days to go back}

\item{end_date}{end date (default = Sys.time())}
}
\value{
tibble
}
\description{
Get historical prices from Yahoo Finance
}
