% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{yahoo_price_for_date}
\alias{yahoo_price_for_date}
\title{Get price for a specific date}
\usage{
yahoo_price_for_date(symbol, date, verbose = FALSE)
}
\arguments{
\item{symbol}{symbol}

\item{date}{date}
}
\description{
Get price for a specific date
}
