% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{yahoo_search}
\alias{yahoo_search}
\title{Search Yahoo Finance for tickers}
\usage{
yahoo_search(q)
}
\arguments{
\item{q}{search query (for example, ISIN, ticker symbol)}
}
\value{
tibble
}
\description{
Search Yahoo Finance for tickers
}
