% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{yahoo_summary}
\alias{yahoo_summary}
\title{Get Yahoo Finance summary data}
\usage{
yahoo_summary(
  symbols,
  modules = c("defaultKeyStatistics", "financialData", "price", "quoteType",
    "summaryDetail"),
  verbose = FALSE
)
}
\arguments{
\item{symbols}{ticker symbols}

\item{modules}{Yahoo Finance modules}
}
\value{
tibble
}
\description{
Get Yahoo Finance summary data
}
\examples{
yahoo_summary(c("AAPL", "MSFT"))
}
